% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{age_codify}
\alias{age_codify}
\title{age_codify}
\usage{
age_codify(.data, .channel, .bird_age)
}
\arguments{
\item{.data}{A data frame or data frame extension (e.g., tibble).}

\item{.channel}{A DBI object that connects to the internal NEFSC Oracle database.}

\item{.bird_age}{Name of the column field that identifies the bird age.}
}
\value{
The mutated tibble.
}
\description{
Replace bird age with code from database.
}
