% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{association_codify}
\alias{association_codify}
\title{association_codify}
\usage{
association_codify(.data, .channel, .seabird_association)
}
\arguments{
\item{.data}{A data frame or data frame extension (e.g., tibble).}

\item{.channel}{A DBI object that connects to the internal NEFSC Oracle database.}

\item{.seabird_association}{Name of the column field that identifies the seabird association.}
}
\value{
The mutated tibble.
}
\description{
Replace seabird association with code from database.
}
