% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{behavior_codify}
\alias{behavior_codify}
\title{behavior_codify}
\usage{
behavior_codify(.data, .channel, .seabird_behavior)
}
\arguments{
\item{.data}{A data frame or data frame extension (e.g., tibble).}

\item{.channel}{A DBI object that connects to the internal NEFSC Oracle database.}

\item{.seabird_behavior}{Name of the column field that identifies the seabird behavior.}
}
\value{
The mutated tibble.
}
\description{
Replace seabird behavior with code from database.
}
