% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{datetime_stampify}
\alias{datetime_stampify}
\title{datetime_stampify}
\usage{
datetime_stampify(.data, .datetime, .tz = "UTC")
}
\arguments{
\item{.data}{A data frame or data frame extension (e.g., tibble).}

\item{.datetime}{Name of the column field that identifies the datetime.}

\item{.tz}{Time zone of the datetime.}
}
\value{
The mutated tibble with a new column labeled datetime_[.tz].
}
\description{
Create a datetime column from character strings.
Assumes datetime has a timezone of UTC, but this can be modified with the .tz argument.
}
