% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{direction_codify}
\alias{direction_codify}
\title{direction_codify}
\usage{
direction_codify(.data, .flight_direction)
}
\arguments{
\item{.data}{A data frame or data frame extension (e.g., tibble).}

\item{.flight_direction}{Name of the column field that identifies the flight direction.}
}
\value{
The mutated tibble.
}
\description{
Replace flight direction with code from database.
Flight direction should be in degrees and is to be entered relative to the ship.
}
