% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{height_codify}
\alias{height_codify}
\title{height_codify}
\usage{
height_codify(.data, .channel, .flight_height)
}
\arguments{
\item{.data}{A data frame or data frame extension (e.g., tibble).}

\item{.channel}{A DBI object that connects to the internal NEFSC Oracle database.}

\item{.flight_height}{Name of the column field that identifies the flight height.}
}
\value{
The mutated tibble.
}
\description{
Replace flight height with code from database.
}
