% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{observer_codify}
\alias{observer_codify}
\title{observer_codify}
\usage{
observer_codify(.data, .channel, .observer_name)
}
\arguments{
\item{.data}{A data frame or data frame extension (e.g., tibble).}

\item{.channel}{A DBI object that connects to the internal NEFSC Oracle database.}

\item{.observer_name}{Name of the column field that identifies the observer name.}
}
\value{
The mutated tibble.
}
\description{
Replace observer name with code from database.
}
