% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{plumage_codify}
\alias{plumage_codify}
\title{plumage_codify}
\usage{
plumage_codify(.data, .channel, .seabird_plumage)
}
\arguments{
\item{.data}{A data frame or data frame extension (e.g., tibble).}

\item{.channel}{A DBI object that connects to the internal NEFSC Oracle database.}

\item{.seabird_plumage}{Name of the column field that identifies the seabird plumage.}
}
\value{
The mutated tibble.
}
\description{
Replace seabird plumage with code from database.
}
