% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{remove_apostrophe}
\alias{remove_apostrophe}
\title{remove_apostrophe}
\usage{
remove_apostrophe(.data, ...)
}
\arguments{
\item{.data}{A data frame or data frame extension (e.g., tibble).}

\item{...}{Name of the column(s) to remove apostrophes from.}
}
\value{
The mutated tibble.
}
\description{
Remove apostrophes from column field.
}
