% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{sex_codify}
\alias{sex_codify}
\title{sex_codify}
\usage{
sex_codify(.data, .sex)
}
\arguments{
\item{.data}{A data frame or data frame extension (e.g., tibble).}

\item{.sex}{Name of the column field that identifies the seabird sex.}
}
\value{
The mutated tibble.
}
\description{
Replace seabird sex with code from database.
}
