% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{upload_data_csv}
\alias{upload_data_csv}
\title{upload_data_csv}
\usage{
upload_data_csv(.directory_path, ...)
}
\arguments{
\item{...}{Arguments to be passed to readr::read_csv.}

\item{directory_path}{A directory path string.}
}
\value{
The concatenated tibble with column names in snake case.
}
\description{
Read in csv data files and concatenate.
}
