% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{upload_data_excel}
\alias{upload_data_excel}
\title{upload_data_excel}
\usage{
upload_data_excel(.directory_path, ...)
}
\arguments{
\item{.directory_path}{A directory path string.}

\item{...}{Arguments to be passed to readxl::read_excel.}
}
\value{
The concatenated tibble with column names in snake case.
}
\description{
Read in excel data files and concatenate.
}
